+++ perf_tests/ze_peer/src/ze_peer_ipc.cpp
@@ -65,7 +65,7 @@ int ZePeer::recvmsg_fd(int socket) {
   return fd;
 }
 
-void ZePeer::set_up_ipc(int number_buffer_elements, uint32_t device_id,
+void ZePeer::set_up_ipc(size_t number_buffer_elements, uint32_t device_id,
                         size_t &buffer_size,
                         ze_command_queue_handle_t &command_queue,
                         ze_command_list_handle_t &command_list) {
@@ -90,7 +90,7 @@ void ZePeer::set_up_ipc(int number_buffer_elements, uint32_t device_id,
 void ZePeer::bandwidth_latency_ipc(peer_test_t test_type,
                                    peer_transfer_t transfer_type,
                                    bool is_server, int commSocket,
-                                   int number_buffer_elements,
+                                   size_t number_buffer_elements,
                                    uint32_t local_device_id,
                                    uint32_t remote_device_id) {
