+++ perf_tests/ze_peer/src/ze_peer_unidirectional.cpp
@@ -45,7 +45,7 @@ void ZePeer::perform_copy(peer_test_t test_type,
 
 void ZePeer::bandwidth_latency(peer_test_t test_type,
                                peer_transfer_t transfer_type,
-                               int number_buffer_elements,
+                               size_t number_buffer_elements,
                                uint32_t remote_device_id,
                                uint32_t local_device_id, uint32_t queue_index) {
